requirements.txt
alembic==1.15.2
altgraph==0.17.4
APScheduler==3.11.0
attrs==25.3.0
autopep8==2.3.2
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
colorama==0.4.6
-e git+https://github.com/thiaguiarfideles/PortaTech.git@0b8cd7800616b7cf03f39cc96336a240013f8cf9#egg=condominio
dlib==19.24.6
easyocr==1.7.2
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-Caching==2.3.1
flask-cors==5.0.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.2
greenlet==3.1.1
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mistune==3.1.3
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
pyclipper==1.3.0.post6
pycodestyle==2.13.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.2
shapely==2.1.0
six==1.17.0
SQLAlchemy==2.0.40
sympy==1.13.1
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
